Generate colour palette from file using colour quantization algorithm
To generate a colour palette, the median cut algorithm is used for colour quantization.
The algorithm selects the subgroup that contains a dimension with the largest range.
This subgroup is divided into two from its median, and this process is repeatedly
done until there are n (currently 5) subgroups. The values in the subgroups are
averaged to get the colours that represent the colour palette.